[PW-6441] move payment request building to order-independent service
